fix: some more layout
