+++ src/Features/Reactflow/components/Decision/DecisionEdge/DecisionEdge.module.scss
@@ -25,7 +25,14 @@
   pointer-events: all;
 }
 
-:global [data-mode="editor"] #task-marker {
+/*
+  We use the WorkflowEngineMode.Edit value
+  to select the task marker
+  There isn't an easier way to do this imo bc
+  the mark is a reused symbol and we only want it to
+  have color in the edit mode
+*/
+:global [data-mode="edit"] #decision-marker {
   polyline {
     fill: var(--flow-switch-primary);
     stroke: var(--flow-switch-primary);
