+++ src/Features/Reactflow/components/End/EndNode.tsx
@@ -9,7 +9,7 @@ export default function EndNode(props: NodeProps) {
   const { mode } = useWorkflowContext();
   const { isConnectable } = props;
   return (
-    <div className={cx(styles.node, { [styles.locked]: mode !== WorkflowEngineMode.Editor })}>
+    <div className={cx(styles.node, { [styles.locked]: mode !== WorkflowEngineMode.Edit })}>
       <Handle
         className={styles.port}
         isConnectable={isConnectable}
