+++ src/Features/Reactflow/components/Start/StartNode.module.scss
@@ -2,7 +2,7 @@
 
 .node {
   background: #002b50;
-  border: 0.125rem solid #161616;
+  border: 0.125rem solid #002b50;
   color: white;
   width: 9rem;
   height: 5rem;
