+++ src/Features/Reactflow/components/Template/TemplateEdge/TemplateEdge.module.scss
@@ -25,7 +25,14 @@
   pointer-events: all;
 }
 
-:global [data-mode="editor"] #task-marker {
+/*
+  We use the WorkflowEngineMode.Edit value
+  to select the task marker
+  There isn't an easier way to do this imo bc
+  the mark is a reused symbol and we only want it to
+  have color in the edit mode
+*/
+:global [data-mode="edit"] #task-marker {
   polyline {
     fill: var(--cds-interactive-01);
     stroke: var(--cds-interactive-01);
