--- src/Features/Reactflow/components/Template/TemplateNode/BaseNode/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './BaseNode';
