+++ src/Features/WorkflowRun/WorkflowRun.module.scss
@@ -49,4 +49,5 @@
   left: 0;
   position: absolute;
   width: 100%;
+  z-index: 1;
 }
